% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Mean-center a vector}
\usage{
center(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector centered on x's mean that is computed with NA values stripped.
}
\description{
Mean-center a vector
}
\examples{
x <- rnorm(193, 100, 15)
center(x)
}
