% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depress2.R
\docType{data}
\name{depress2}
\alias{depress2}
\title{Data sets from the D research}
\format{
\subsection{\code{depress2}}{

A data frame with 173 rows and 104 columns at wave 2:
\describe{
\item{depr02}{Depression Scale, item 02. Item 02, 05, 06, 11, 12, 14, 16, 17, 18,
and 20 has NOT been reverse-coded in \code{depress2}.}
}
}
}
\source{
Keng package.
}
\usage{
depress2
}
\description{
D refers to Depression. D research is a three-wave longitudinal survey research.
Here are four data sets, \code{depress1} only includes data collected at wave 1,
\code{depress2} at wave 2, and \code{depress3} at wave 3. \code{depress} was created by joining \code{depress1},
\code{depress2}, and \code{depress3} together. \code{depress} has done reverse-coding and
has computed variable scores like \code{depr1}, which indicates the depression score at wave 1.
\code{depress} serves to ease users' practicing plotting and statistical analyzing. \code{depress1},
\code{depress2}, and \code{depress3} serves to support users' practicing data wrangling.
}
\keyword{datasets}
