% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide a vector into three groups}
\usage{
divide(x, criterion = "1sd")
}
\arguments{
\item{x}{A numeric vector.}

\item{criterion}{A character value \code{"1sd"} or a numeric value between 0 and 0.5.}
}
\value{
An ordered factor with levels "Low < Medium < High".
}
\description{
Divide a vector into three groups
}
\details{
\code{divide()} divides a vector into three groups using the criterion of 1 SD,
or proportions like 0.27. Differing from \code{cut()}, the medium interval of \code{divide()}
does not include the left and right limits.
}
\examples{
x <- 1:100
divide(x)
divide(x, 0.27)
}
