% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_O.R
\name{move_O}
\alias{move_O}
\title{Move the origin of x vector forward by \code{coord} distance.}
\usage{
move_O(x, coord = NULL)
}
\arguments{
\item{x}{The original vector.}

\item{coord}{Move origin of x vector forward by \code{coord} distance.
If not specified, \code{coord} is set be the mean of x by default.}
}
\value{
A vector whose coordinate centered on x's mean, which is computed with NA values stripped.
}
\description{
Move the origin of x vector forward by \code{coord} distance.
}
\details{
O in the \code{move_O} is capital.
The basic use of \code{move_O} is to mean-center a vector, with the argument \code{coord} set to be NULL by default.
If \code{coord} is not NULL, \code{move_O} will move origin of x by coord distance; in this case,
\code{move_O} could be used to pick points in simple-slope analysis for moderation models.
}
\examples{
x <- rnorm(193, 100, 15)
mean(x)
head(move_O(x)) - head(x)

x <- rnorm(193, 4, 1)
w <- rnorm(193, 3, 2)
y <- 0.2*x + 0.2*w + 0.2*x*w + rnorm(193)
dat <- data.frame(y, x, w)
fit <- lm(y ~ x*w, dat)
coefs <- coef(fit)
# simple slope of x at w = 3
coefs[2] + coefs[4]*3 # 0.7783861
w_3 <- move_O(w, 3)
dat$w_3 <- w_3
fit <- lm(y ~ x*w_3, dat)
coef(fit)[2]
fit <- lm(y ~ x*I(move_O(w, 3)), dat)
coef(fit)[2]
}
