% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_dcb.R
\name{pick_dcb}
\alias{pick_dcb}
\title{Pick Double Color Balls}
\usage{
pick_dcb(size = 1L, verbose = TRUE)
}
\arguments{
\item{size}{The size of sets of Super Lotto numbers to pick.}

\item{verbose}{A logical value. Print the numbers picked or not.}
}
\value{
Print the numbers picked, and return the invisible \code{balls} list that stored these numbers.
}
\description{
Pick Double Color Balls
}
\examples{
pick_dcb(10)
out <- pick_dcb(10, verbose = FALSE)
out
}
