% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Keng_power.R
\name{plot.Keng_power}
\alias{plot.Keng_power}
\title{Plot the power against sample size for the Keng_power class}
\usage{
\method{plot}{Keng_power}(x, ...)
}
\arguments{
\item{x}{The output object of \code{power_r()} or \code{power_lm()}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A plot of power against sample size.
}
\description{
Plot the power against sample size for the Keng_power class
}
\examples{
plot(power_lm())
out <- power_r(0.2, n = 193)
plot(out)
}
