% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE.R
\name{PRE}
\alias{PRE}
\title{PRE}
\usage{
PRE(fitc, fitA)
}
\arguments{
\item{fitc}{The result of \code{lm()} of the Compact model (Model C).}

\item{fitA}{The result of \code{lm()} of the Augmented model (Model A).}
}
\value{
A data.frame including \emph{SSEC}, \emph{SSEA}, \emph{PRE}, \emph{F}, \emph{df1}, \emph{df2}, and \emph{p}. SSEC  = The sum of squared errors of Model C, SSEA  = The sum of squared errors of Model A, PRE = The proportional reduction in error. The rest is the F test of PRE.
}
\description{
PRE
}
\examples{
x <- rnorm(193)
y <- 0.3 + 0.2*x + rnorm(193)
dat <- data.frame(y, x)
fitC <- lm(y ~ 1, dat)
fitA <- lm(y ~ x, dat)
PRE(fitC, fitA)
}
