% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Keng_power.R
\name{print.Keng_power}
\alias{print.Keng_power}
\title{Print the contents of the Keng_power class briefly}
\usage{
\method{print}{Keng_power}(x, ...)
}
\arguments{
\item{x}{The output object of \code{power_r()} or \code{power_lm()}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
None (invisible NULL).
}
\description{
Print the contents of the Keng_power class briefly
}
\examples{
print(power_lm(n = 200))

out <- power_r(0.2, n = 193)

out
}
