% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale.R
\name{Scale}
\alias{Scale}
\title{Scale a vector}
\usage{
Scale(x, m = 0, sd = NULL)
}
\arguments{
\item{x}{The original vector.}

\item{m}{The expected Mean of the scaled vector.}

\item{sd}{The expected Standard Deviation (unit) of the scaled vector.}
}
\value{
The scaled vector.
}
\description{
Scale a vector
}
\details{
To scale \code{x}, its mean, or standard deviation (unit), or both, could be changed.
The missing values of \code{x} are removed if any.

If \code{m} = 0 (default), and \code{sd} = \code{NULL}(default), \code{x} would be mean-centered (default).

If \code{m} = \code{NULL}, the mean of \code{x} would not changed.
If \code{m} is not \code{NULL}, the mean of \code{x} would be transformed to \code{sd}.

If \code{m} = \code{NULL} or 0, and \code{sd} = 1, \code{x} would be standardized to be its z-score with m = 0 and sd = 1.
The mean and standardized score are not necessarily 0 and 1.
}
\examples{
(x <- rnorm(10, 5, 2))
# Mean-center x.
Scale(x)
Scale(x, m = 0)
# Transform the mean of x to 3.
Scale(x, m = 3)
# Transform x to its z-score.
Scale(x, sd = 1)
# Standardize x with m = 100 and sd = 15.
Scale(x, m = 100, sd = 15)
}
