% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\title{Standardize a vector}
\usage{
std(x, M = 0, SD = 1)
}
\arguments{
\item{x}{The original vector.}

\item{M}{The Mean of the standardized vector.}

\item{SD}{The Standard Deviation of the standardized vector.}
}
\value{
The standardized vector.
}
\description{
Standardize a vector
}
\details{
The standardized score is not necessarily the z-score.
The \emph{M} and \emph{SD} of the standardized vector should be set according to your need.
For example, the \emph{M} and \emph{SD} of IQ are 100 and 15, respectively.
To standardization x, the \emph{M} and \emph{SD} of x are needed and computed,
for which the missing values of x are removed if any.
}
\examples{
x <- rnorm(193)
IQ <- std(x, 100, 15)
}
