% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_r.R
\name{test_r}
\alias{test_r}
\title{Test \emph{r} using t-test given \emph{r} and \emph{n}.}
\usage{
test_r(r, n)
}
\arguments{
\item{r}{Pearson correlation.}

\item{n}{sample size of \emph{r}.}
}
\value{
\emph{r}, \emph{se} of \emph{r}, \emph{t}, and \emph{p}.
}
\description{
Test \emph{r} using t-test given \emph{r} and \emph{n}.
}
\examples{
test_r(0.2, 193)
}
