% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyY2XYZ.R
\name{xyY2XYZ}
\alias{xyY2XYZ}
\title{Transform CIE xyY c(x, y, Y) to matrix(c(X, Y, Z))}
\usage{
xyY2XYZ(x, y, Y)
}
\arguments{
\item{x}{CIE coordinate. An atomic numeric value or a numeric vector.}

\item{y}{CIE coordinate. An atomic numeric value or a numeric vector.}

\item{Y}{CIE coordinate. An atomic numeric value or a numeric vector.}
}
\value{
A 3-column matrix whose rows specify the colors, and cols specify the XYZs.
}
\description{
Transform CIE xyY c(x, y, Y) to matrix(c(X, Y, Z))
}
\examples{
XYZ <- xyY2XYZ(0.555, 0.328, 9.4)
RGB <- convertColor(XYZ, "XYZ", "sRGB")
rgb(RGB[,1], RGB[,2], RGB[,3])
}
