% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_r.R
\name{cut_r}
\alias{cut_r}
\title{Cut-off values of r when sample size n is known}
\usage{
cut_r(n)
}
\arguments{
\item{n}{Sample size of the r}
}
\value{
Cut-off values of r at the significance levels of p = 0.1, 0.05, 0.01, 0.001.  r with the absolute value larger than the cut-off value is significant at the corresponding significance level.
}
\description{
Cut-off values of r when sample size n is known
}
\examples{
cut_r(193)
}
