% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_r.R
\name{test_r}
\alias{test_r}
\title{Test r using t-test when r and n is known}
\usage{
test_r(r, n)
}
\arguments{
\item{r}{Pearson correlation}

\item{n}{sample size of r}
}
\value{
r, se of r, t, and p.
}
\description{
Test r using t-test when r and n is known
}
\examples{
test_r(0.2, 193)
}
